:recycle: refactor: main.py를 제거하고 클래스명을 변경한다
