+++ check_latest_lotto_result.py
@@ -3,8 +3,9 @@ from textwrap import dedent
 from typing import NamedTuple
 
 from lotto.account import Account, fetch_account
+from lotto.lotto import Lotto
+from lotto.lotto_site import LottoSite
 from lotto.secret import Secret
-from main import Lotto, LottoWebsite
 from sends.send import Send, SendResult
 from sends.send_github_issue import SendGithubIssue
 
@@ -41,7 +42,7 @@ def check_latest_lotto_result(account: Account, lotto: Lotto, send: Send, search
 if __name__ == '__main__':
     check_latest_lotto_result(
         account=fetch_account(),
-        lotto=LottoWebsite(),
+        lotto=LottoSite(),
         send=SendGithubIssue(token=Secret('ghp_'), repository='viiviii/jubilant-train'),  # todo: 하드코딩
         search_dates=(last_sunday(date.today()), date.today())  # todo: 일요일~토요일로 변경하기
     )
