+++ lotto/account.py
@@ -3,26 +3,14 @@ from typing import Optional
 
 import keyring
 
+from lotto.lotto import Account
+
 ID_ARGUMENT_NAME = '--id'
 ID_ARGUMENT_OPTIONS = {'type': str, 'help': '로또 사이트 계정 아이디'}
 KEYRING_SERVICE_NAME = 'lotto-purchase-keyring'
 
 
-class Account(object):
-    def __init__(self, account_id: str, account_password: str) -> None:
-        self.id = account_id
-        self.password = account_password
-
-    def __str__(self) -> str:
-        def mask(string) -> str:
-            return '*' * len(string)
-
-        return f'id={self.id}, ' \
-               f'password={mask(self.password) if self.password else self.password}'
-
-    def __repr__(self) -> str:
-        return f'Account({self})'
-
+# todo: 파일명 바꿔라
 
 def fetch_account() -> Account:
     _id = _id_from_args()
