+++ tests/conftest.py
@@ -1,8 +1,7 @@
 import pytest
 from _pytest.config.argparsing import Parser
 
-from lotto import account as lotto_account
-from lotto.account import Account, fetch_account
+from lotto.account import Account, fetch_account, ID_ARGUMENT_OPTIONS, ID_ARGUMENT_NAME
 
 
 @pytest.fixture(scope='session')
@@ -15,5 +14,4 @@ def pytest_addoption(parser):
 
 
 def _add_id_option_for_passing_argument_to_pytest(parser: Parser):
-    parser.addoption(lotto_account.ID_ARGUMENT_NAME,
-                     **lotto_account.ID_ARGUMENT_OPTIONS)
+    parser.addoption(ID_ARGUMENT_NAME, **ID_ARGUMENT_OPTIONS)
